Third test passed for invalid id delete

+++ src/main/java/com/learn/project/thanksgiving/Entity/Item.java
@@ -9,7 +9,7 @@ import javax.persistence.Id;
 public class Item {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     private String name;
 
@@ -17,16 +17,16 @@ public class Item {
 
     }
 
-    public Item(int id, String name) {
+    public Item(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
